backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow>=2.10.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
